+++ src/main/java/DaoOfModding/Cultivationcraft/Common/Qi/BodyParts/FoodStats/QiFoodStats.java
@@ -259,7 +259,7 @@ public class QiFoodStats extends FoodData
 
     // Is called when meditating
     // Is supplied the amount of Qi the player can absorb during this turn, and returns any modifications to that amount
-    public int meditation(int QiRemaining, Player player)
+    public double meditation(double QiRemaining, Player player)
     {
         return QiRemaining;
     }
